proposal type moved to separate module, all its methods moved to separate module to be more goway
