+++ chains/evm/calls/bridge.go
@@ -1,17 +1,20 @@
 package calls
 
 import (
+	"context"
 	"fmt"
-	"github.com/ChainSafe/chainbridge-core/chains/evm/voter"
+	"github.com/ChainSafe/chainbridge-core/chains/evm/voter/proposal"
 	"math/big"
 	"strings"
 
 	"github.com/ChainSafe/chainbridge-core/chains/evm/calls/consts"
-
+	"github.com/ChainSafe/chainbridge-core/relayer"
+	"github.com/ethereum/go-ethereum"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/rs/zerolog/log"
 )
+
 
 func PrepareSetBurnableInput(handler, tokenAddress common.Address) ([]byte, error) {
 	a, err := abi.JSON(strings.NewReader(consts.BridgeABI))
@@ -131,7 +134,7 @@ func Deposit(client ChainClient, fabric TxFabric, bridgeAddress, recipient commo
 	return nil
 }
 
-func ExecuteProposal(client ClientDispatcher, fabric TxFabric, proposal *voter.Proposal) (common.Hash, error) {
+func ExecuteProposal(client ClientDispatcher, fabric TxFabric, proposal *proposal.Proposal) (common.Hash, error) {
 	// revertOnFail should be constantly false, true is used only for internal contract calls when you need to execute proposal in voteProposal function right after it becomes Passed becouse of votes
 	input, err := PrepareExecuteProposalInput(proposal.Source, proposal.DepositNonce, proposal.ResourceId, proposal.Data, true)
 	if err != nil {
@@ -146,7 +149,7 @@ func ExecuteProposal(client ClientDispatcher, fabric TxFabric, proposal *voter.P
 }
 
 
-func VoteProposal(client ClientDispatcher, fabric TxFabric, proposal *voter.Proposal) (common.Hash, error) {
+func VoteProposal(client ClientDispatcher, fabric TxFabric, proposal *proposal.Proposal) (common.Hash, error) {
 	// revertOnFail should be constantly false, true is used only for internal contract calls when you need to execute proposal in voteProposal function right after it becomes Passed becouse of votes
 	input, err := PrepareVoteProposalInput(proposal.Source, proposal.ResourceId, proposal.Data)
 	if err != nil {
@@ -170,4 +173,68 @@ func PrepareSetDepositNonceInput(domainID uint8, depositNonce uint64) ([]byte, e
 		return []byte{}, err
 	}
 	return input, nil
+}
+
+type ContractCallerClient interface {
+	CallContract(ctx context.Context, callArgs map[string]interface{}, blockNumber *big.Int) ([]byte, error)
+}
+
+func ProposalStatus(evmCaller ContractCallerClient, p *proposal.Proposal) (relayer.ProposalStatus, error) {
+	a, err := abi.JSON(strings.NewReader(consts.BridgeABI))
+	if err != nil {
+		return relayer.ProposalStatusActive, err
+	}
+	input, err := a.Pack("getProposal", p.Source, p.DepositNonce, p.Data)
+	if err != nil {
+		return relayer.ProposalStatusInactive, err
+	}
+
+	msg := ethereum.CallMsg{From: common.Address{}, To: &p.BridgeAddress, Data: input}
+	out, err := evmCaller.CallContract(context.TODO(), ToCallArg(msg), nil)
+	if err != nil {
+		return relayer.ProposalStatusInactive, err
+	}
+	type bridgeProposal struct {
+		ResourceID    [32]byte
+		DataHash      [32]byte
+		YesVotes      []common.Address
+		NoVotes       []common.Address
+		Status        uint8
+		ProposedBlock *big.Int
+	}
+	res, err := a.Unpack("getProposal", out)
+	if err != nil {
+		return relayer.ProposalStatusInactive, err
+	}
+	out0 := *abi.ConvertType(res[0], new(bridgeProposal)).(*bridgeProposal)
+	return relayer.ProposalStatus(out0.Status), nil
+}
+
+func idAndNonce(srcId uint8, nonce uint64) *big.Int {
+	var data []byte
+	data = append(data, big.NewInt(int64(nonce)).Bytes()...)
+	data = append(data, uint8(srcId))
+	return big.NewInt(0).SetBytes(data)
+}
+
+func ProposalVotedBy(evmCaller ContractCallerClient, by common.Address, p *proposal.Proposal) (bool, error) {
+	a, err := abi.JSON(strings.NewReader(consts.BridgeABI))
+	if err != nil {
+		return false, err
+	}
+	input, err := a.Pack("_hasVotedOnProposal", idAndNonce(p.Source, p.DepositNonce), p.GetDataHash(), by)
+	if err != nil {
+		return false, err
+	}
+	msg := ethereum.CallMsg{From: common.Address{}, To: &p.BridgeAddress, Data: input}
+	out, err := evmCaller.CallContract(context.TODO(), ToCallArg(msg), nil)
+	if err != nil {
+		return false, err
+	}
+	res, err := a.Unpack("_hasVotedOnProposal", out)
+	if err != nil {
+		return false, err
+	}
+	out0 := *abi.ConvertType(res[0], new(bool)).(*bool)
+	return out0, nil
 }
