+++ chains/evm/cli/bridge/register-resource.go
@@ -3,6 +3,7 @@ package bridge
 import (
 	"encoding/hex"
 	"fmt"
+
 	"github.com/ChainSafe/chainbridge-core/chains/evm/calls"
 	"github.com/ChainSafe/chainbridge-core/chains/evm/cli/flags"
 	"github.com/ChainSafe/chainbridge-core/chains/evm/evmclient"
