--- chains/evm/voter/proposal.go
@@ -1,113 +1,0 @@
-package voter
-
-import (
-	"context"
-	"math/big"
-	"strings"
-
-	"github.com/ethereum/go-ethereum/crypto"
-	"github.com/ChainSafe/chainbridge-core/chains/evm/calls"
-	"github.com/ChainSafe/chainbridge-core/relayer"
-	"github.com/ethereum/go-ethereum"
-	"github.com/ethereum/go-ethereum/accounts/abi"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/rs/zerolog/log"
-)
-
-func NewProposal(source uint8, depositNonce uint64, resourceId [32]byte, data []byte, handlerAddress, bridgeAddress common.Address) *Proposal {
-	return &Proposal{
-		Source:         source,
-		DepositNonce:   depositNonce,
-		ResourceId:     resourceId,
-		Data:           data,
-		HandlerAddress: handlerAddress,
-		BridgeAddress:  bridgeAddress,
-	}
-}
-
-type Proposal struct {
-	Source         uint8  // Source domainID where message was initiated
-	DepositNonce   uint64 // Nonce for the deposit
-	ResourceId     [32]byte
-	Payload        []interface{} // data associated with event sequence
-	Data           []byte
-	HandlerAddress common.Address
-	BridgeAddress  common.Address
-}
-
-func (p *Proposal) Status(evmCaller ChainClient) (relayer.ProposalStatus, error) {
-	definition := "[{\"inputs\":[{\"internalType\":\"uint8\",\"name\":\"originDomainID\",\"type\":\"uint8\"},{\"internalType\":\"uint64\",\"name\":\"depositNonce\",\"type\":\"uint64\"},{\"internalType\":\"bytes32\",\"name\":\"dataHash\",\"type\":\"bytes32\"}],\"name\":\"getProposal\",\"outputs\":[{\"components\":[{\"internalType\":\"bytes32\",\"name\":\"_resourceID\",\"type\":\"bytes32\"},{\"internalType\":\"bytes32\",\"name\":\"_dataHash\",\"type\":\"bytes32\"},{\"internalType\":\"address[]\",\"name\":\"_yesVotes\",\"type\":\"address[]\"},{\"internalType\":\"address[]\",\"name\":\"_noVotes\",\"type\":\"address[]\"},{\"internalType\":\"enumBridge.ProposalStatus\",\"name\":\"_status\",\"type\":\"uint8\"},{\"internalType\":\"uint256\",\"name\":\"_proposedBlock\",\"type\":\"uint256\"}],\"internalType\":\"structBridge.Proposal\",\"name\":\"\",\"type\":\"tuple\"}],\"stateMutability\":\"view\",\"type\":\"function\"}]"
-	a, err := abi.JSON(strings.NewReader(definition))
-	if err != nil {
-		return relayer.ProposalStatusInactive, err // Not sure what status to use here
-	}
-	input, err := a.Pack("getProposal", p.Source, p.DepositNonce, p.Data)
-	if err != nil {
-		return relayer.ProposalStatusInactive, err
-	}
-
-	msg := ethereum.CallMsg{From: common.Address{}, To: &p.BridgeAddress, Data: input}
-	out, err := evmCaller.CallContract(context.TODO(), toCallArg(msg), nil)
-	if err != nil {
-		return relayer.ProposalStatusInactive, err
-	}
-	type bridgeProposal struct {
-		ResourceID    [32]byte
-		DataHash      [32]byte
-		YesVotes      []common.Address
-		NoVotes       []common.Address
-		Status        uint8
-		ProposedBlock *big.Int
-	}
-	res, err := a.Unpack("getProposal", out)
-	if err != nil {
-		return relayer.ProposalStatusInactive, err
-	}
-	out0 := *abi.ConvertType(res[0], new(bridgeProposal)).(*bridgeProposal)
-	return relayer.ProposalStatus(out0.Status), nil
-}
-
-func (p *Proposal) VotedBy(evmCaller ChainClient, by common.Address) (bool, error) {
-	definition := "[{\"inputs\":[{\"internalType\":\"uint72\",\"name\":\"\",\"type\":\"uint72\"},{\"internalType\":\"bytes32\",\"name\":\"\",\"type\":\"bytes32\"},{\"internalType\":\"address\",\"name\":\"\",\"type\":\"address\"}],\"name\":\"_hasVotedOnProposal\",\"outputs\":[{\"internalType\":\"bool\",\"name\":\"\",\"type\":\"bool\"}],\"stateMutability\":\"view\",\"type\":\"function\"}]"
-	a, err := abi.JSON(strings.NewReader(definition))
-	if err != nil {
-		return false, err // Not sure what status to use here
-	}
-	input, err := a.Pack("_hasVotedOnProposal", idAndNonce(p.Source, p.DepositNonce), p.GetDataHash(), by)
-	if err != nil {
-		return false, err
-	}
-	msg := ethereum.CallMsg{From: common.Address{}, To: &p.BridgeAddress, Data: input}
-	out, err := evmCaller.CallContract(context.TODO(), toCallArg(msg), nil)
-	if err != nil {
-		return false, err
-	}
-	res, err := a.Unpack("_hasVotedOnProposal", out)
-	if err != nil {
-		return false, err
-	}
-	out0 := *abi.ConvertType(res[0], new(bool)).(*bool)
-	return out0, nil
-}
-
-func (p *Proposal) Vote(client ChainClient, fabric calls.TxFabric) error {
-	hash, err := calls.VoteProposal(client, fabric, p)
-	if err != nil {
-		return err
-	}
-	log.Debug().Str("hash", hash.String()).Uint64("nonce", p.DepositNonce).Msgf("Voted")
-	return nil
-}
-
-// GetDataHash constructs and returns proposal data hash
-func (p *Proposal) GetDataHash() common.Hash {
-	return crypto.Keccak256Hash(append(p.HandlerAddress.Bytes(), p.Data...))
-}
-
-
-func idAndNonce(srcId uint8, nonce uint64) *big.Int {
-	var data []byte
-	data = append(data, big.NewInt(int64(nonce)).Bytes()...)
-	data = append(data, uint8(srcId))
-	return big.NewInt(0).SetBytes(data)
-}
