+++ chains/evm/voter/voter.go
@@ -6,11 +6,13 @@ package voter
 import (
 	"context"
 	"fmt"
+	"github.com/ChainSafe/chainbridge-core/chains/evm/voter/proposal"
 	"math/big"
 
 	"github.com/ChainSafe/chainbridge-core/chains/evm/calls"
 	"github.com/ChainSafe/chainbridge-core/relayer"
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/rs/zerolog/log"
 )
 
 type ChainClient interface {
@@ -21,18 +23,11 @@ type ChainClient interface {
 	calls.ClientDispatcher
 }
 
-type Proposer interface {
-	Status(client ChainClient) (relayer.ProposalStatus, error)
-	VotedBy(client ChainClient, by common.Address) (bool, error)
-	Vote(client ChainClient, fabric calls.TxFabric) error
-}
-
 type MessageHandler interface {
-	HandleMessage(m *relayer.Message) (Proposer, error)
+	HandleMessage(m *relayer.Message) (*proposal.Proposal, error)
 }
 
 type EVMVoter struct {
-	stop   <-chan struct{}
 	mh     MessageHandler
 	client ChainClient
 	fabric calls.TxFabric
@@ -51,18 +46,19 @@ func (w *EVMVoter) VoteProposal(m *relayer.Message) error {
 	if err != nil {
 		return err
 	}
-	ps, err := prop.Status(w.client)
+	ps, err := calls.ProposalStatus(w.client, prop)
 	if err != nil {
 		return fmt.Errorf("error getting proposal: %+v status %w", prop, err)
 	}
-	votedByTheRelayer, err := prop.VotedBy(w.client, w.client.RelayerAddress())
+	votedByTheRelayer, err := calls.ProposalVotedBy(w.client, w.client.RelayerAddress(), prop)
 	if err != nil {
 		return err
 	}
 	// if this relayer had not voted for proposal and proposal in Active status then we need to vote for
 	// And that basically it o other options compared to previous contracts version
 	if !votedByTheRelayer && ps == relayer.ProposalStatusActive {
-		err = prop.Vote(w.client, w.fabric)
+		hash, err := calls.VoteProposal(w.client, w.fabric, prop)
+		log.Debug().Str("hash", hash.String()).Uint64("nonce", prop.DepositNonce).Msgf("Voted")
 		if err != nil {
 			return fmt.Errorf("Voting failed. Err: %w", err)
 		}
